View and preview on speaker console now try to use the publics ratio of width to height.
Added multilanguage support. Now german as well as english available. Language depends on the browsers language setting. Default remains english.

Function nextStep changed to manage more komplex html documents where the html elements may be in differnt branches of the DOM tree.

In cases when the speaker console can not be opend a button overlay is showen top open the windows by this button. This can help when Firefox prevents a not interactive windows.open.