Make impressConsole.js a self contained impress.js plugin
This patch makes necessary changes to adhere to the new impress.js
plugin api, as described here:
https://github.com/henrikingo/impress.js/blob/myfork/src/plugins/README.md

See https://github.com/regebro/impress-console/issues/19 for a thorough
discussion on what that means for impressConsole in particular. This
patch implements the changes described in that issue.

Even if you don't care about such a plugin api, this introduces the following
enhancements for any user of impress-console:

 * impressConsole will now call init() on itself automatically when
   you call impress().init(). This means you only need to include
   impressConsole.js file with a <script>, and it will work out of
   the box.

 * The default css is inlined into impressConsole.js. This means that
   impressConsole.js is now a self contained file and doesn't depend
   on any other file (e.g. the css files) for working.

 * All changes are backward compatible. For example, you can still call
   `impressConsole().init('css/impressConsoleDark.css')` if you want.